+++ oap-server/server-query-plugin/promql-plugin/src/main/java/org/apache/skywalking/oap/query/promql/rt/exception/ParseErrorListener.java
@@ -24,8 +24,6 @@ import org.antlr.v4.runtime.RecognitionException;
 import org.antlr.v4.runtime.Recognizer;
 
 public class ParseErrorListener extends BaseErrorListener {
-    public static final ParseErrorListener INSTANCE = new ParseErrorListener();
-
     @Override
     public void syntaxError(Recognizer<?, ?> recognizer,
                             Object offendingSymbol,